pipelines/optuna-pipeline/optuna-worker/pyproject.toml
[tool.poetry]
name = "optuna-worker"
version = "1.0"
description = ""
authors = ["tonouchi510 <tonouchi27@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9,<3.10"
setuptools = "^57.4.0"
google-cloud-storage = "^1.42.0"
PyYAML = "5.4.1"
optuna = "^2.10.0"
absl-py = ">=0.9,<=0.11"
google-cloud-secret-manager = "^2.8.0"
kfp = "^1.8.10"
pandas = "^1.3.5"
PyMySQL = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"